@@ -3,11 +3,11 @@ package guru.springframework.jdbc.dao;
 import guru.springframework.jdbc.domain.Author;
 import org.springframework.stereotype.Component;
 
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.TypedQuery;
-import javax.persistence.criteria.*;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.TypedQuery;
+import jakarta.persistence.Query;
+import jakarta.persistence.criteria.*;
 import java.util.List;
 
 /**
