@@ -1,6 +1,6 @@
 package guru.springframework.jdbc.domain;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.Objects;
 
 @NamedQueries({
